requirements.txt
anthropic>=0.40.0
crewai>=0.86.0
chromadb>=0.5.0
sqlalchemy>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
streamlit>=1.40.0
plotly>=5.18.0
pandas>=2.1.0
python-dotenv>=1.0.0
